requirements_dev.txt
pip
holidays
bump2version
wheel
watchdog
flake8
tox
coverage
Sphinx
twine
Click
pytest
pytest-runner
dateutils
